requirements.txt
Flask==3.1.0
Flask_CKEditor==1.0.0
Flask_Login==0.6.3
icecream==2.1.3
markdown2==2.5.1
python-dotenv==1.0.1
requests==2.28.1
Werkzeug==3.1.3
